requirements.txt
Fabric
Flask-SQLAlchemy
Flask-WTF
WTForms
coverage
Flask-Login==0.4.0
# ecdsa
# paramiko
# pycrypto
